Start handling click events from the user
Plumbed a click handler down to the individual squares, and started to
flesh it out in App.js. It currently cycles the current shape, but not
much else.

Also refactored the sub-components of InfoBar and Board to be instance
methods instead of separate functions - I think this ends up being a
much cleaner design, and probably closer to what the React creators
intended.

There's currently a bug where sometimes the current shape will render
partially or not at all. No idea why.